requirements.txt
google-generativeai==0.7.2
pillow==10.3.0
streamlit== 1.35.0
jsonpatch==1.33
jsonpointer== 2.4
jsonschema == 4.21.1
